Fix MSVC warning C4267 by casting to the correct type explictly

+++ include/libtorrent/bencode.hpp
@@ -212,7 +212,7 @@ namespace libtorrent
 				break;
 			case entry::preformatted_t:
 				std::copy(e.preformatted().begin(), e.preformatted().end(), out);
-				ret += e.preformatted().size();
+				ret += static_cast<int>(e.preformatted().size());
 				break;
 			case entry::undefined_t:
